chore: rename package of test example
