+++ app/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.saradar.showloading"
+        applicationId "com.saradar.customloadertest"
         minSdkVersion 21
         targetSdkVersion 28
         versionCode 1
