+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">ShowLoading</string>
+    <string name="app_name">CustomLoaderTest</string>
 </resources>
